Make a PEP 561 package
Henceforth, this should be used as a Python package rather than as a
submodule. This (I think) did not exist at the time we originally
created this repo, and this method is broadly preferred. It removes the
need for (oft maligned) Git submodules, and also makes it
straightforward to add our own package of “real” Python code alongside
that adds types that are not upstream, removing the ceremony needed to
monkey-patch them in within the implementation code.