airflow: Add type for Context
This is rather opaque and difficult to find documentation for. This is
but a subset of what it has, but it’s for the most part what we use
often.

Full(ish) documentation is (I think) at
https://airflow.apache.org/docs/apache-airflow/stable/templates-ref.html
but of course Airflow docs make no reference to “context” there and it’s
frustratingly difficult to actually find docs for what is in the
context. Hence the desire to type this to stop guessing.

[OHF-852]