airflow: Fix most models being implicitly Any
Since most models inherit from Base, and Base was Any, that meant that
basically any undefined property on any instance of (a descendent of)
Base would not be a type error. This remedies that at the cost of
potentially missing some things in Base, but practically it’s likely not
relevant to us as downstream consumers rather than developing Airflow
itself.